+++ CHANGELOG.md
@@ -9,9 +9,11 @@ with the exception that 0.x versions can break between minor versions.
 ## [Unreleased]
 ### Added
 - Add `RegexOptionsBuilder::build_delegate_prefilter` to disable building the prefilter
-- Add `ByteSet` that returns the first bytes that can match a pattern so you can build your own prefiler
+- Add `RegexOptionsBuilder::start_bytes` that returns the first bytes that can match a pattern so you can build your own prefilter
+- Add `RegexOptionsBuilder::required_bytes` that returns the required bytes for a pattern to match
 
 ### Changed
+- If a lookbehind consists of an Alt containing only easy expressions and at least 4 branches (and no capture groups), compile it as a variable lookbehind (if the feature is enabled) instead of trying every branch as its own const-size lookbehind. One reverse-DFA delegate over the whole alternation is much cheaper per attempt, and grants a perf boost.
 ### Fixed
 
 ## [0.19.2] - 2026-09-13
@@ -19,6 +21,8 @@ with the exception that 0.x versions can break between minor versions.
 - Add a `leftmost_longest` feature (off by default) that adds a method to the `RegexBuilder`/`RegexOptionsBuilder` for building a VM which would operate in leftmost-longest match mode instead of the regular leftmost-first mode. Useful for POSIX compliance. (#281)
 ### Changed
 ### Fixed
+- Fix a panic in `capture_names()` for a pattern with a zero-repeated named group such as `(?<n>a){0}` (#275)
+- Literal bytes were being encoded as utf-8 in Ascii mode, now they match the exact byte literal
 
 ## [0.19.1] - 2026-09-06
 ### Added